+++ src/App.js
@@ -2,7 +2,7 @@ import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import MyProfile from './components/MyProfile';
 import Missions from './components/Missions';
-import Rockets from './components/Rockets';
+import Rockets from './components/Rockets/Rockets';
 import Dragons from './components/Dragons';
 
 function App() {
