+++ src/index.css
@@ -1,3 +1,5 @@
-.link:active {
-  text-decoration: none;
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
